Make html escaping available to derive
