+++ rinja_derive/src/html.rs
@@ -1,0 +1,1 @@
+../../rinja/src/html.rs
